refactor: change armor weight to 0 and update weapon attack speed to …

+++ src/data/Data.cpp
@@ -21,7 +21,7 @@ namespace WeaponData {
             int cost = weapon["cost"];
             WeaponType type = utils::stringToWeaponType(weapon["type"]);
             int damage = weapon["damage"];
-            int atkSpeed = weapon["atkSpeed"];
+            float atkSpeed = weapon["atkSpeed"];
             _availableWeapons.push_back(Weapon(name, cost, type, damage, atkSpeed));
         }
     }
